refactor(2B): relocate AdbcReader into the CDK sql package (M3)
The ADBC-only source reader is stdlib + pyarrow only (no engine
imports), so it belongs in the CDK where the GenericSQLConnector's read
path will consume it. git mv preserves history; the source connector and
sql package re-export are rewired, and the pure-reader execution tests
move to the CDK test tree while the connector-branch wiring tests stay
engine-side.
